update kết nối dish và topping

+++ src/main/java/com/isp392/repository/DishRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface DishRepository extends JpaRepository<Dish, Long> {
+public interface DishRepository extends JpaRepository<Dish, Integer> {
     boolean existsByDishName(String dishName);
     Optional<Dish> findByDishName(String dishName);
 }
